Split secure tunnel and mqtt tuneel manager
