+++ beluga-ssh-service/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "beluga-tunnel"
+name = "beluga-ssh-service"
 version = "0.1.0"
 authors.workspace = true
 edition.workspace = true
@@ -10,14 +10,7 @@ publish.workspace = true
 
 [dependencies]
 bytes.workspace = true
-enumn.workspace = true
-prost.workspace = true
-futures.workspace = true
-serde.workspace = true
-serde_json.workspace = true
-thiserror.workspace = true
-tracing.workspace = true
-tokio-tungstenite = { version = "0.21", features = ["rustls-tls-webpki-roots"] }
+beluga-tunnel = { path = "../beluga-tunnel" }
 
 [dependencies.tokio]
 workspace = true
